Add a protocol for agents
Add a `Protocol` subclass to define the agent protocol[1] (aka the
interface).

This also adds two other modules.

`chat`: the place for chat-related type and value definitions. Right
now, it is just re-exporting the OpenAI chat types. By making our
package use the type definitions from this module, we can later easily
swap out the OpenAI types to more flexible types that work with
different providers.

`logging`: holds a logger that we can use throughout the package, and
exposes that logger to the package users so they can modify the logger
(log-level, etcetera).

[1]: https://typing.readthedocs.io/en/latest/spec/protocol.html