Add a simple memory protocol and class (#11)
Add a simple memory protocol that dictates the basic interface for
acting on agent memory.

Add a `WithMemory` class that acts as memory storage. The user is
supposed to use this as a mix-in to an `BaseAgent` implementation, which
can then make use of memory. Although, technically they could also
interact directly with the memory class.