Make tweaks to the OpenAI agent (#14)
Make some tweaks to the OpenAI agent class:

It inherits (and thus conforms to) the `BaseAgent` protocol. In doing
so, we implemented the `count_tokens` method and the `create_completion`
method.

`create_completion` basically does the same thing as the the behavior of
calling `OpenAI().chat.completions.create(...)` via the `__getattr__`
passthrough. However, it does make requests to various pre and post
processor functions, which the passthrough to the underlying OpenAI
client does not do. We should reconcile that later.

We also added some new features to the `OpenAI` agent, which are passed
in during initialization and otherwise occur in the background.

You can init it with `pre_completion_fns`, which are called before the
inference request. They accept the messages as an arg, and return a
modified messages list (or they can do something else).

You can init the object with `completion_callbacks`, which are called
after the inference request succeeds. They accept both the messages and
the resulting completion.

You can init the object `memory`, which maintains a memory of all past
LLM calls.

Also, squeeze in a couple fixes for the tests, making the
`openai_test.py` use pytest assertions.