Make Fixpoint ChatCompletion cleanly wrap inner completion (#23)
Make the Fixpoint ChatCompletion cleanly wrap the inner completion. We
make our implementation a subclass of the OpenAI `ChatCompletion` so
that we get proper type-checking for free.

I also moved the file from `src/fixpoint/completions/fixpoint.py` to
`src/fixpoint/completions/chat_completion.py` and we don't prefix it
with `FixpointChatCompletion` because that's kind of a naming stutter.

Within the code-base, I replaced all direct external usage of the OpenAI
`ChatCompletion` class with our internal `ChatCompletion` class.