Hook up workflows into agent
Create a `fixpoint.workflow` module and then hook up the workflow
objects into an agent's `create_completion` method, so that we can mark
the completion as belonging to that workflow.

We modified the `fixpoint.memory` so that each memory item is now a
`MemoryItem` class instead of a tuple of input messages and output
completion. Then we could add the workflow entry to the `MemoryItem`.