Reconcile OpenAI interface and our BaseAgent protocol (#24)
We have a `BaseAgent` protocol, which is an API that is not constrained
to be OpenAI compliant, which is part of the core of the Fixpoint
framework.

However, we also want an OpenAI-compliant API, so we have two different
classes for interacting with OpenAI:

- `OpenAI` - intended to be a drop-in replacement for the
  `openai.OpenAI` client.
- `OpenAIAgent` - interacts with the OpenAI API, but conforms to our
  `BaseAgent` protocol.

When we make a completion from the `OpenAIAgent`, it can interact with
memory, and pre and post-completion hooks. The
`agent.chat.completions.create` function used to not have any of these
features, but now we have reconciled the two so that the `OpenAI` class
uses the `OpenAIAgent` under the hood so we get these extra goodies.