Allow control over when agent caches
Allow us to control when an agent caches. We add a `CacheMode` parameter
to the agent implementation classes, and also let the caller of
`agent.create_completion` manually specify a cache mode override.

Also add some debug logging throughout the `fixpoint.cache` module.
There is a `fixpoint.cache.logger` object so that you can control the
log level for the caching, specifically.