Fix caching key stability bug
The Python `hash` function is not stable across new instances of the
Python VM. Furthermore, with some caches it would be useful to preserve
the semantic meaning of the cache keys instead of randomizing them. For
example, if we want to later add support to iterate over all cache key
entries or do semantic caching.

To solve both these problems, we stop hashing with the builtin `hash`
function to pseudo-random integers, and make these two changes:

1. The `DiskTLRUCache` requires you pass in a `serialize_key_fn` that
   converts the key to a string
2. The `DiskTLRUCache` doesn't do any special stuff, instead just using
   the implementations default pickling for serialization.