ChatCompletion has generic type for structured output
Change the `ChatCompletion` class definition so that it accepts a
generic type that can be any subclass of `pydantic.BaseModel`. This is
the type of the structured data that the `ChatCompletion` can optionally
hold.

Then I had to wire up the generics across the codebase when using chat
completions. It would be nice if we could make the generic type optional
by giving it a default type of `BaseModel`, but we'll need to wait for
PEP-696 for that.[1]

One downside of doing this, is whenever we call `create_completion`
without specifying a `response_model` argument, we need to annotate the
return type with `ChatCompletion[BaseModel]`. This is kind of annoying,
and necessary until we have PEP-696.

[1]: https://peps.python.org/pep-0696/