Allow caching of Pydantic models inside ChatCompletions (#38)
When we try to cache a ChatCompletion to disk, if it has an inner
"structured output" Pydantic model, then the `DiskCache` throws an error
when trying to pickle the object. It's possible this error is just
related to slightly different import paths on cache save versus load,
but regardless, saving pickled objects is not a good strategy.

So we add support for serializing a ChatCompletion to disk as JSON, and
loading it back in as JSON. When we load it back in as JSON, we don't
know what `Pydantic` class to use for the structured output, so we
extended the caching classes to accept a Pydantic class when calling
`get`. This was added to a new protocol, `SupportsChatCompletionCache`.

Most of the rest of this commit is just fixing all of the plumbing for
caching to use the new `SupportsChatCompletionCache` and pass in the
Pydantic module to the cache when available.

Later, we should probably consider making private the generic
`SupportsCache` and its generic class implementations. There is no need
to expose them publicly, and we don't want to use them anyways, when
interacting with LLM inference, since we always use a cache that is
specifically for input/output inference messages.