Remove `SupportsWorkflow` and just use `Workflow`
We had `SupportsWorkflow` which was wrapping a simple dataclass
`Workflow`. It was kind of verbose. Just use `Workflow` for now. I don't
think we'll have different workflow implementations, but we will have
different storage/durability layers for workflows. That is not a concern
of a workflow class, but rather the storage layer, which can either
process a workflow or be injected into the workflow when constructing
it.

Also, make the `Workflow` a Pydantic model for easier serialization.