Improve workflow form serialization and updates (#63)
Make the Workflow Form serialization/deserialization logic use the
underlying JSON schema. This is more durable for long-term storage than
relying on import paths of Pydantic models, and theoretically it lets us
support non-Python languages in the future.

As part of this, we limit the types of forms that users can create so
that it is easy to do form deserialization. We force the form to:

1. only support simple non-nested fields: bool, str, int, float
2. each field must be optional, since the form starts empty and we fill
   it in

We also update the logic for how a form is updated. When you call
`form.update`, it only updates the fields you specify, leaving the other
ones as is.

Finally, make a little change in how Supabase storage does form
serialization and deserialization. We used to prefer to serialize by
calling Pydantic's `BaseModel.model_dump`, but now we prefer to use
`serialize`/`deserialize` if they exist, because we have more control
over them.