Make chat completion caching support more chat completion parameters (#65)
When we make a chat completion request, we used to only do caching based
on the "messages" parameter. However, there are a bunch of other model
parameters that affect the response, such as the model used or the
temperature.

Now, we cache based on more completion request parameters. We don't
cover all arguments to OpenAI yet, but more of them.

Along the way, I tried to make serialization and deserialization of
work for pairs of:

- CreateChatCompletionRequest
- ChatCompletion

The problem here is that both of these types contain a Pydantic model
class and instance, respectively. This type does not serialize or
deserialize cleanly. You must know what class to use when deserializing.

We kind of kludge this into working for now. It was difficult for me to
use the `TLRUCache`. I tried to get that working, but kept running into
problems. So instead of the `ChatCompletionTLRUCache` inheriting from
the `TLRUCache`, it is now a totally separate class.

For the most part, it serialization and deserialization now works, both
for the cache and for Supabase storage.