Fix a few bugs in structured workflow example
Fix a few bugs in the structured workflow example so that it runs. Also,
use `TaskGroup` for structured concurrency to wait on multiple
coroutines. Structured concurrency is a more readable pattern for
knowing when all concurrent calls are finished.

Make a few updates to how we represent the workflow node-state path-like
ID string. If we are in the main task or step (aka no task or step
specified), then we omit that part from the path ID.

When you store a document or form, by default we store it at the
workflow's current node.