Structured workflows (#64)
This is an example of the API interface for structured workflows. We
want structured workflows to function in "library mode", where someone
is just running them on their local machine for local development or
testing. We also want them to be able to run in "infra mode', where they
integrate with a workflow platform, likely Temporal.

Right now, it doesn't do any checkpointing or properly update the
workflow context and/or workflow run to indicate what task or step we
are on.

The main goals of this PR were to:

- Define the API interface for structured workflows
- The interface should be compatible with Temporal, ultimately
- Make sure that we can actually execute the code in the interface, ie
  run a workflow
- Validate the format of structured workflow definitions
- Set up the skeleton so that we can easily add hooks and attributes to
  a workflow, task, or step

From here, we can actually implement keeping track of workflow context
and doing workflow state checkpointing, etc.