Move imperative workflow tests (#69)
Move the imperative workflow tests under `tests/extras` which matches up
with the source path.

Also, update schema for storing Supabase migrations. We use a schema,
`fpmeta`, that is separate from the `fixpoint` schema, because otherwise
the migrations get busted when we try to create or delete the `fixpoint`
schema inside a migration.