Set up factories for workflow storage
Set up factories for configuring all necessary storage for workflows.
This includes:

- form storage
- document storage
- agent memory
- agent cache

Also, create a Postgres migration file that configures the appropriate
tables. I took the example tables from within the tests, but I made some
changes to them. Generally, I made some columns non-nullable, added
workflow info to a few tables, and made the memory table support
multiple agents. I also updated tests so that the adhoc postgres table
creations use the new table formats.

As part of making memory work for multiple agents, we need each agent to
have a ID. Right now, you can specify a ID when creating an agent, and
if you don't it will be given a random ID.

When you create a structured workflow, you pass in a dictionary of agent
IDs to agents, and we can set the agent names based on that, if they
were originally randomly set. We don't do that yet, though.