[structured] Extend WorkflowRunHandle functionality
Extend the `WorkflowRunHandle` with methods to check the status of the
workflow (open/closed) and to get the inner `WorkflowRun` object via
`WorkflowRunHandle.finalized_workflow_run(...)`.
`finalized_workflow_run` will only return a `WorkflowRun` if we finished
awaiting the coroutine result of the workflow.

Also, squeeze in a little more code into the `examples/structured_workflow.py`
file that shows how we can save/load the chat completion results to a
dataframe.