Add caching layer for task and step calls
Add a caching layer for when we call tasks and steps so that we can
check if we already called them before.

The caching layer is not hooked up yet to task and step definitions, but
we have the new `workflows/structured/_callcache.py` module that defines
the cache class and we added a `decorate_with_cache` function to the
`_helpers.py` module.

Remaining work:

- make a client for persistent caching across executions
- hook it up to the task and step decorators
- (later) support disk caching and Postgres caching