Rework setup code for workflows and agents (#72)
Rework some setup code for workflows and agents. Changes:

**Agent and memory changes**

Agents now have a `memory` field, which is how you can refer to an
agent's memory. Before, you would pass in a memory container but there
was no way to access the memory container.

Memory is required on all agents. If you don't supply a memory
container, we'll use a no-op memory container that doesn't actually
store memories. This makes it easier for the rest of the code to assume
that we are storing memories.

I had to move the `_memgpt.py` file so that we no longer had a circular
import, which was causing a Pylint error.

**Workflow changes**

The `WorkflowContext` is no longer a dataclass, so we have control over
its initialization.

`WorkflowContext` no longer accepts a memory field. Instead, the agents
all have memory containers themselves.