Add callcache converter file for deserialization to types (#108)
Add a `_converter.py` file to the `structured._callcache` module. The
`_converter` module is for deserializing text into typed values. Without
this, we would just deserialize JSON to simple dicts or lists or
something, but with this we can deserialize based on type-hints, thus
supporting Pydantic or dataclass deserialization.

This was taken from Temporal, which we attribute at the top of the
source file.