Fix bug when cloning wrapped workflow agents
The `structured/_context.py:WorkflowContext` class inherits from the
imperative `WorkflowContext`, but they had different `__init__`
signatures. So when we called `clone` (defined on the imperative
`WorkflowContext`) it was using the structured context `__init__`
method, which was missing a keyword argument.

Fix this by adding the missing argument to the structured
`WorkflowContext.__init__` method.