Add ability to retry failing workflow runs
If a WorkflowRun fails due to a transient issue, Fixpoint can retry the
WorkflowRun. This adds support for retries in both the imperative and
the structural workflows.

**Imperative workflows**

A `WorkflowRun` now has an `attempt_id` for every attempt to complete
the run. These are the two added fields for `WorkflowRun`:

- `attempt_id(self) -> str`
- `generate_new_attempt(self) -> str` - mutates the attempt ID for the
  `WorkflowRun`

On the `Workflow` object, you can call `Workflow.retry(run_id)` to try
loading in an existing workflow run and beginning a new attempt.

**Structured workflows**

In the `workflows.structured` package, we add `retry_workflow` and
`respawn_workflow`.

`retry_workflow` is semantically similar to `run_workflow` but you must
pass in a run ID. `respawn_workflow` is the equivalent for
`spawn_workflow`.

Because run ID's are system generated, there are two ways to get a run
ID for a workflow. The simplest, is if you call `spawn_workflow` you can
pull it off the `WorkflowRunHandle`. If you are using `run_workflow`,
you can wrap your call to `run_workflow` in an exception handler for an
`ExecutionException`, which has a `workflow_run_id` field.

**Error changes**

All (or almost all) Fixpoint-defined error classes inherit from the
`FixpointException` in the `fixpoint.errors` module.