Documents can be stored outside workflow runs or workflows
Change the API and storage layers for documents so that you can store a
document rooted either:

1. on a workflow and a workflow run (the only way previously supported)
2. on a workflow, but outside of any workflow run
3. outside of any workflows

We had to change the storage layers to use composite primary keys for
documents. Of note, you cannot have any column of a primary key be
nullable in Postgres, so if a document's `workflow_id` or
`workflow_run_id` column is NULL, we actually change it to the special
string `__null__`.

Fixes Linear issue: https://linear.app/fixpoint/issue/PRO-60/convert-primary-keys-for-docs-and-forms-to-allow-duplicate-ids-in